GISTWithTarget

General Iterative Shrinkage and Thresholding Algorithm based on Gong, P., Zhang, C., Lu, Z., Huang, J., & Ye, J. (2013). A General Iterative Shrinkage and Thresholding Algorithm for Non-convex Regularized Optimization Problems. In S. Dasgupta & D. McAllester (Eds.), Proceedings of Machine Learning Research (PMLR; Vol. 28, Issue 2, pp. 37--45). PMLR. http://proceedings.mlr.press

Usage

GISTWithTarget(
  cpptsemObject,
  startingValues,
  objective,
  lambda,
  adaptiveLassoWeights,
  regularizedParameters,
  targetVector,
  eta,
  sig,
  initialStepsize,
  stepsizeMin,
  stepsizeMax,
  GISTLinesearchCriterion,
  GISTNonMonotoneNBack,
  maxIter_out,
  maxIter_in,
  break_outer,
  verbose,
  silent
)

Arguments

cpptsemObject

model of type MxObject to compute the gradients

startingValues

named vector with starting values

objective

"ML" for maximum likelihood SEM, "Kalman" for Kalman filter

lambda

penalty value

adaptiveLassoWeights

named vector with adaptive lasso weights

regularizedParameters

named vector of regularized parameters

targetVector

named vector with values towards which the parameters are regularized

eta

if the current step size fails, eta will decrease the step size. Must be > 1

sig

Controls the sigma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421, Equation 20. Defaults to 0. Has to be in 0 < sigma < 1

initialStepsize

initial stepsize to be tried in the outer iteration

stepsizeMin

Minimal acceptable step size. Must be > 0. A larger number corresponds to a smaller step from one to the next iteration. All step sizes will be computed as described by Gong et al. (2013)

stepsizeMax

Maximal acceptable step size. Must be > stepsizeMin. A larger number corresponds to a smaller step from one to the next iteration. All step sizes will be computed as described by Gong et al. (2013)

GISTLinesearchCriterion

criterion for accepting a step. Possible are 'monotone' which enforces a monotone decrease in the objective function or 'non-monotone' which also accepts some increase.

GISTNonMonotoneNBack

in case of non-monotone line search: Number of preceding regM2LL values to consider

maxIter_out

maximal number of outer iterations

maxIter_in

maximal number of inner iterations

break_outer

Stopping criterion for outer iterations. It has to be a named value. By default (name: gradient), a relative first-order condition is checked, where the maximum absolute value of the gradients is compared to break_outer (see https://de.mathworks.com/help/optim/ug/first-order-optimality-measure.html). Alternatively, an absolute tolerance can be passed to the function (e.g., break_outer = c("gradient" = .0001)). Instead of relative gradients, the change in parameters can used as breaking criterion. To this end, use c("parameterChange" = .00001)

verbose

set to 1 to print additional information and plot the convergence and 2 for further details.

silent

suppress all warning messages

Details

GIST minimizes a function of form f(theta) = l(theta) + g(theta), where l is the likelihood and g is a penalty function. Various penalties are supported, however currently only lasso and adaptive lasso are implemented.

NOTE: Function located in file GIST.R