approx_iterateOverLambdas

loops over lambdas if optimization = "approx"

Usage

approx_iterateOverLambdas(
  cpptsemObject,
  dataset,
  sampleSize,
  regIndicators,
  lambdas,
  penalty,
  adaptiveLassoWeights,
  targetVector,
  returnFitIndices,
  BICWithNAndT,
  Tpoints,
  objective,
  epsilon,
  zeroThresh,
  controlApproxOptimizer,
  scaleLambdaWithN,
  verbose
)

Arguments

cpptsemObject

Fitted object of type cpptsem

dataset

wide data set

sampleSize

sample size

regIndicators

matrix with ones and zeros specifying which parameters in regOn should be regularized. Must be of same size as the regularized matrix. 1 = regularized, 0 = not regularized. Alternatively, labels for the regularized parameters can be used (e.g. drift_eta1_eta2)

lambdas

vector of penalty values (tuning parameter). E.g., seq(0,1,.01)

penalty

type. Currently supported are lasso and ridge for optimization = approx and lasso for optimization = exact

adaptiveLassoWeights

weights for the adaptive lasso.

targetVector

named vector with values towards which the parameters are regularized (Standard is regularization towards zero)

returnFitIndices

Boolean: should fit indices be returned?

BICWithNAndT

Boolean: TRUE = Use N and T in the formula for the BIC (-2log L + log(N+T)*k, where k is the number of parameters in the model). FALSE = Use both N in the formula for the BIC (-2log L + log(N))

Tpoints

Number of time points (used for BICWithNAndT)

objective

which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)

epsilon

epsilon is used to transform the non-differentiable lasso penalty to a differentiable one if optimization = approx

zeroThresh

threshold below which parameters will be evaluated as == 0 in lasso regularization if optimization = approx

controlApproxOptimizer

settings passed to optimx or Rsolnp

scaleLambdaWithN

Boolean: Should the penalty value be scaled with the sample size? True is recommended, as the likelihood is also sample size dependent

verbose

0 (default), 1 for convergence plot, 2 for parameter convergence plot and line search progress

Details

NOTE: Function located in file approx_optimization.R

Author

Jannik Orzek