exact_bfgsGLMNET

Performs GLMNET (see Friedman, 2010 & Yuan, 2011) with bfgs approximated Hessian

Usage

exact_bfgsGLMNET(
  cpptsemObject,
  dataset,
  objective,
  regIndicators,
  lambdas,
  adaptiveLassoWeights,
  sparseParameters,
  stepSize,
  lineSearch,
  c1,
  c2,
  sig,
  gam,
  initialHessianApproximation,
  maxIter_out,
  maxIter_in,
  maxIter_line,
  eps_out,
  eps_in,
  eps_WW,
  scaleLambdaWithN,
  sampleSize,
  approxFirst,
  numStart,
  controlApproxOptimizer,
  extraTries,
  verbose
)

Arguments

cpptsemObject

Fitted object of class cpptsem

dataset

only required if objective = "Kalman". Please provide a data set in wide format compatible to ctsemOMX

objective

which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)

regIndicators

Vector with names of regularized parameters

lambdas

Vector with lambda values that should be tried

adaptiveLassoWeights

weights for the adaptive lasso.

sparseParameters

labeled vector with parameter estimates of the most sparse model.

stepSize

Initial stepSize of the outer iteration (theta_k+1 = oldParameters + stepSize * Stepdirection)

lineSearch

String indicating which linesearch should be used. Defaults to the one described in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421. Alternatively (not recommended) Wolfe conditions (lineSearch = "Wolfe") can be used in the outer iteration. Setting to "none" is also not recommended!.

c1

c1 constant for lineSearch. This constant controls the Armijo condition in lineSearch if lineSearch = "Wolfe"

c2

c2 constant for lineSearch. This constant controls the Curvature condition in lineSearch if lineSearch = "Wolfe"

sig

only relevant when lineSearch = 'GLMNET'. Controls the sigma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421.

gam

Controls the gamma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421. Defaults to 0.

initialHessianApproximation

Which initial hessian approximation should be used? Possible are: 'ident' for an identity matrix and 'OpenMx' (here the hessian approxmiation from the mxObject is used). If the Hessian from 'OpenMx' is not positive definite, the negative Eigenvalues will be 'flipped' to positive Eigenvalues. This works sometimes, but not always. Alternatively, a matrix can be provided which will be used as initial Hessian

maxIter_out

Maximal number of outer iterations

maxIter_in

Maximal number of inner iterations

maxIter_line

Maximal number of iterations for the lineSearch procedure

eps_out

Stopping criterion for outer iterations

eps_in

Stopping criterion for inner iterations

eps_WW

Stopping criterion for weak Wolfe line search. If the upper - lower bound of the interval is < epsWW, line search will be stopped and stepSize will be returned

scaleLambdaWithN

Boolean: Should the penalty value be scaled with the sample size? True is recommended, as the likelihood is also sample size dependent

sampleSize

sample size for scaling lambda with N

approxFirst

Should approximate optimization be used first to obtain start values for exact optimization?

numStart

Used if approxFirst = 3. regCtsem will try numStart+2 starting values (+2 because it will always try the current best and the parameters provided in sparseParameters)

controlApproxOptimizer

settings passed to optimx or Rsolnp

extraTries

number of extra tries in mxTryHard for warm start

verbose

0 (default), 1 for convergence plot, 2 for parameter convergence plot and line search progress

Details

NOTE: Function located in file GLMNET.R